--- SeniorSolutionsWeb/Controllers/LoginController.cs
@@ -1,46 +1,0 @@
-﻿using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using System.Security.Claims;
-
-namespace SeniorSolutionsWeb.Controllers
-{
-    public class LoginController : Controller
-    {
-        public IActionResult Index(string returnUrl)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            return View();
-        }
-        [HttpPost("Login")]
-        public async Task<IActionResult> ValidateUser(string username, string password, string returnUrl)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            if(username == "admin" && password == "admin")
-            {
-                var claims = new List<Claim>();
-                claims.Add(new Claim("username", username));
-                claims.Add(new Claim(ClaimTypes.NameIdentifier, username));
-                var claimsIdentity = new ClaimsIdentity(claims, CookieAuthenticationDefaults.AuthenticationScheme);
-                var claimsPrincipal = new ClaimsPrincipal(claimsIdentity);
-                await HttpContext.SignInAsync(claimsPrincipal);
-                return Redirect(returnUrl);
-            }
-            TempData["Error"] = "Username or password is invalid.";
-            return View("Index");
-        }
-
-        public async Task<IActionResult> LogoutUser()
-        {
-            await HttpContext.SignOutAsync();
-            return RedirectToAction("Login");
-        }
-
-        [Authorize]
-        public IActionResult Secured()
-        {
-            return View();
-        }
-    }
-}
