+++ SeniorSolutionsWeb/Views/Account/AccessDenied.cshtml
@@ -1,0 +1,8 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
+@{
+    ViewData["Title"] = "Unauthorized";
+}
+<h1>Unauthorized User</h1>
+<p>We're sorry, you are not allowed to view the page you requested. Please contact your supervisor.</p>
