+++ SeniorSolutionsWeb/Views/Account/Login.cshtml
@@ -1,0 +1,23 @@
+﻿@{
+    ViewData["Title"] = "User Login Portal";
+    var returnUrl = ViewData["ReturnUrl"] as string;
+    var error = TempData["Error"] as string;
+    Layout = "_NoNavbar";
+}
+<h3>@error</h3>
+<div class="col-12 d-flex justify-content-center align-items-center vh-100">
+    <div class="col-4 p-5 bg-white border border-5 rounded">
+        <form action="/login?ReturnUrl=@System.Net.WebUtility.UrlEncode(returnUrl)" method="post">
+            <div class="mb-3">
+                <label for="username" class="form-label fw-bold">Username</label>
+                <input type="text" class="form-control" id="username" name="username">
+            </div>
+            <div class="mb-3">
+                <label for="password" class="form-label fw-bold">Password</label>
+                <input type="password" class="form-control" id="password" name="password">
+            </div>
+            <button type="submit" class="btn btn-primary">Submit</button>
+        </form>
+    </div> 
+    
+</div>
