+++ SeniorSolutionsWeb/Views/Account/Secured.cshtml
@@ -1,0 +1,14 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
+@{
+}
+<h1>This is a secure page.</h1>
+<h2>Welcome @User.Identity.Name</h2>
+
+<ul>
+    @foreach (var i in @User.Claims)
+    {
+        <li>@i.Type @i.Value</li>
+    }
+</ul>
