--- SeniorSolutionsWeb/Views/Login/Index.cshtml
@@ -1,20 +1,0 @@
-﻿@{
-    ViewData["Title"] = "login";
-    var returnUrl = ViewData["ReturnUrl"] as string;
-    var error = TempData["Error"] as string;
-    Layout = "_Layout";
-}
-
-<h1>login</h1>
-<h3>@error</h3>
-<form action="/login?ReturnUrl=@System.Net.WebUtility.UrlEncode(returnUrl)" method="post">
-  <div class="mb-3">
-    <label for="username" class="form-label">Username</label>
-    <input type="text" class="form-control" id="username" name="username">
-  </div>
-  <div class="mb-3">
-    <label for="password" class="form-label">Password</label>
-    <input type="password" class="form-control" id="password" name="password">
-  </div>
-  <button type="submit" class="btn btn-primary">Submit</button>
-</form>
