+++ SeniorSolutionsWeb/Views/Shared/_Layout.cshtml
@@ -32,13 +32,28 @@
                             <a class="nav-link" asp-controller="Manager" asp-action="Index">Manager's Page</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-controller="Login" asp-action="Secured">Secure Page</a>
+                            <a class="nav-link" asp-controller="Account" asp-action="Secured">Secure Page</a>
                         </li>
+                        @if (User.Identity.IsAuthenticated)
+                        {
+                            <li class="nav-item">
+                                <a class="nav-link" asp-controller="Account" asp-action="LogoutUser">Logout</a>
+                            </li>
+                        }
                     </ul>
                 </div>
             </div>
         </nav>
     </header>
+    <div class="container-fluid">
+        <div class="row">
+            <environment exclude="Development">
+                <div class="col-12">
+                    <p class="text-danger fw-bold">Current Environment: Development</p>
+                </div> 
+            </environment>
+        </div>
+    </div>
     <div class="container">
         <main role="main" class="pb-3">
             @RenderBody()
